// src/app/(admin)/apps/file/types.ts
import type { StaticImageData } from 'next/image';

export type QuickAccessItem = {
	icon: string;
	name: string;
	size: string;
};
type Member = {
	image: StaticImageData;
	name: string;
};
export type RecentFileItem = {
	name: string;
	modifiedDate: string;
	modifiedBy: string;
	size: string;
	owner: string;
	members: Member[];
};